--- fnc/math_fnc.py
@@ -1,3 +1,0 @@
-
-def get_kb_str_fm_bytes(len_: int):
-    return f"{len_/1024:.2f} kb"
